### 2024-04-05 #6583 @editorialbot
[PRE REVIEW]: HinetPy: A Python package for accessing and processing NIED Hi-net seismic data
**Submitting author:** @seisman (<a href="http://orcid.org/0000-0001-7967-1197">Dongdong Tian</a>)
**Repository:** https://github.com/seisman/HinetPy
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** 0.8.1
