### 2024-04-05 #6583 @kthyng
@seisman I am going to ping the editorial board to see if this is in scope for JOSS in terms of it being a web client which might not fit. This will take a few weeks, thanks for your patience.
