+++ .github/workflows/docs.yml
@@ -46,7 +46,7 @@ jobs:
         run: make doc
 
       - name: Deploy to GitHub Pages
-        uses: peaceiris/actions-gh-pages@v3.9.3
+        uses: peaceiris/actions-gh-pages@v4.0.0
         # Only deploy on main branch
         if: github.ref == 'refs/heads/main'
         with:
