### 2024-05-24 #6583 @kthyng
Hi @seisman and sorry for the delay in getting back to you. I have a couple of questions.

1. One editor noted: "I notice the paper mentions ObsPy (https://docs.obspy.org/) as a standard package for accessing seismological data. I'd be happier seeing this implemented as an extension to that package, to make the NIED Hi-net data in the paper consistent with the other sources already supported by ObsPy." Is there a possibility of integrating this work into ObsPy fully or more closely, or having it as a sort of extension?
2. I noticed the note in the paper:
> Users must register an account on the NIED Hi-net website for accessing the data and renew the account annually.
Is this account free and automatic or does it need to be manually approved? Would it be possibly to have some test data available long-term for reviewers and for testing purposes that doesn't require an account?
