### 2024-05-25 #6583 @seisman
> Hi @seisman and sorry for the delay in getting back to you. I have a couple of questions.

No worries. Thank you for considering this submission.

> 1. One editor noted: "I notice the paper mentions ObsPy (https://docs.obspy.org/) as a standard package for accessing seismological data. I'd be happier seeing this implemented as an extension to that package, to make the NIED Hi-net data in the paper consistent with the other sources already supported by ObsPy." Is there a possibility of integrating this work into ObsPy fully or more closely, or having it as a sort of extension?

In the seismological community, the [International Federation of Digital Seismograph Networks (FDSN)](https://www.fdsn.org/) defines the [FDSN web service interfaces](https://www.fdsn.org/webservices/) that provide access to data and metadata in seismology. The ObsPy package's [`obspy.clients.fdsn` module](https://docs.obspy.org/packages/obspy.clients.fdsn.html#module-obspy.clients.fdsn) provides a Pythonic interface of the FDSN web service, to access seismic data from data centers as long as the data center's web server implements the FDSN web services. As of today, there are already many data centers (https://www.fdsn.org/datacenters/) that support FDSN web services, but NIED Hi-net still doesn't. Accessing seismic data from NIED Hi-net is different from accessing via FDSN web services, so it's impossible to integrate the HinetPy package into ObsPy.

> 2. I noticed the note in the paper:
>
> > Users must register an account on the NIED Hi-net website for accessing the data and renew the account annually.
>
> Is this account free and automatic or does it need to be manually approved? Would it be possibly to have some test data available long-term for reviewers and for testing purposes that doesn't require an account?

Yes, the account is free. Users can register the account by filling out an online form (https://hinetwww11.bosai.go.jp/nied/registration/?LANG=en) and need to wait for a few business days for approval.

The key feature of the HinetPy package is accessing seismic data from NIED Hi-net, so local test data is insufficient for reviewers to evaluate the package. What's more, the NIED Hi-net prohibits any distribution of their data, so an account is required for reviewing the package.
