### 2024-05-29 #6583 @kthyng
Hi @seisman. This submission has been a bit tricky to decide on since it is both related to a web application and heavily involves data access (both of which aren't always in scope), but we think it is in scope for JOSS so let's move forward.
