### 2024-05-29 #6583 @elbeejay
Hi @seisman, thanks for the submission and your answer above provides some of the information reviewers are going to need to have in order to test and evaluate your software. I'm going to be handling the editorial process for your submission going forward; do you have any potential reviewers in mind that you might suggest? This seems like a good opportunity for others in your field to sign up and access data from NIED Hi-net so it definitely will make sense to identify the people who are in the same field to review this work.

If you do have suggestions, please provide their names and github handles here (omit the "@" so they don't get tagged). Thanks!
