### 2024-06-04 #6583 @stefanazzz
Of course, glad to help with that. I used hynetpy very recently. I wish
there was a python equivalent for fnet , too. Please use my work address
***@***.***
Thanks
Stefan
