### 2024-06-04 #6583 @elbeejay
Hi Stefan - since this is happening through GitHub I believe the notifications are automatically routed through to the email address you have setup with GitHub so I can't control that unfortunately.
