### 2024-06-04 #6840 @editorialbot
[REVIEW]: HinetPy: A Python package for accessing and processing NIED Hi-net seismic data
**Submitting author:** @seisman (<a href="http://orcid.org/0000-0001-7967-1197">Dongdong Tian</a>)
**Repository:** https://github.com/seisman/HinetPy
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v0.9.0
