### 2024-06-05 50e27df
Fix a small bug and update client.py
Fix a small bug:OSError: [Errno 18] Invalid cross-device link.

os.rename only works if source and destination are on the same file system. Use shutil.move instead.
