+++ HinetPy/client.py
@@ -13,6 +13,7 @@ import zipfile
 from datetime import datetime, timedelta
 from html.parser import HTMLParser
 from multiprocessing.pool import ThreadPool
+import shutil
 
 import requests
 from requests.adapters import HTTPAdapter
@@ -524,8 +525,8 @@ class ContinuousWaveformClient(BaseClient):
             ctable = os.path.join(dirname, ctable)
         if dirname and not os.path.exists(dirname):
             os.makedirs(dirname, exist_ok=True)
-        os.rename(ch_euc, ctable)
-
+        #os.rename(ch_euc, ctable)
+        shutil.move(ch_euc, ctable)
         # 4. cleanup
         if cleanup:
             for cnt in cnts:
