### 2024-06-05 01deb1f
Update HinetPy/client.py
