### 2024-06-05 54361a8
Also fix the renaming bug in win32.py
