### 2024-06-06 seisman/HinetPy#117 @YuYifan2000
JOSS Review
Thanks Dr. Tian for preparing and sharing the HinetPy package for the seismology community, it is a useful tool designed to assist downloading the high quality HiNet data which facilitates tremendous scientific findings. When reviewing your submission to JOSS, I found following minor issues that I hope Dr. Tian could resolve:

1. Could you explain briefly in the paper which python packages or techniques are implemented in HinetPy that makes data quires possible.
2. Even though you made it clear in the installation part in the documentation that the users need to install win32tool from NIED, could you also briefly state that the win32 functionality of the HinetPy package is based on the win32tool installed from NIED in the paper?
3. Based on the 'community guideline' of the review checklist, the information about 2) Report issues or problems with the software and 3) Seek support are missing in the repository. Even though the issues are well maintained and answered timely by the authors.

That's all concerns I have for the author. Thanks again for the brilliant and open work.

Yifan Yu
