### 2024-06-07 e5731ed
Revert "Set verify=False"
This reverts commit 4b0bdd644df249a91d06a4d3ac2e1cfcddadee53.
