### 2024-06-07 fce2bbf
Upgrade certifi and explicitly set the certification [skip ci] (#118)
