+++ .github/workflows/tests.yml
@@ -44,6 +44,7 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install -r requirements.txt
+          python -m pip install --upgrade certifi
           python -m pip install build pytest pytest-cov coverage[toml] codecov
           python -m build --sdist
           python -m pip install dist/*
