### 2024-06-07 f0cc9f1
Clarify that requests and win32tools are required
