### 2024-06-07 seisman/HinetPy#117 @YuYifan2000
Thanks for your reply. Your submission meets the checklist for JOSS based on my criteria.
