### 2024-06-14 #6840 @elbeejay
Awesome, thanks @YuYifan2000 and @zhong-yy for your prompt reviews. The rapid comments and revisions between @YuYifan2000 and @seisman in https://github.com/seisman/HinetPy/issues/117 are exactly how the process is intended to work.

@seisman there's one small note from @zhong-yy at the bottom of his review: "The paper is well written except that there is a typo in line 38: 24,000 channels should have been 2,400." I'm just highlighting it here so you don't miss it!
