### 2024-06-14 #6840 @seisman
Thanks @YuYifan2000 and @zhong-yy for the quick reviews.

> The paper is well written except that there is a typo in line 38: 24,000 channels should have been 2,400.

Fixed in https://github.com/seisman/HinetPy/commit/4c974054775c10de507afd2876e0ac5d66dce390.
