+++ .github/workflows/pypi-publish.yml
@@ -29,6 +29,6 @@ jobs:
 
     - name: Publish to PyPI
       if: startsWith(github.ref, 'refs/tags')
-      uses: pypa/gh-action-pypi-publish@v1.8.14
+      uses: pypa/gh-action-pypi-publish@v1.9.0
       with:
         password: ${{ secrets.PYPI_API_TOKEN }}
