### 2024-06-19 de3213c
CI: Improve the Tests workflow
