+++ .github/workflows/tests.yml
@@ -13,6 +13,11 @@ on:
       - 'HinetPy/**'
       - 'tests/**'
       - '.github/workflows/tests.yml'
+  workflow_dispatch:
+
+concurrency:
+  group: ${{ github.workflow }}-${{ github.ref }}
+  cancel-in-progress: ${{ github.ref != 'refs/heads/main' }}
 
 jobs:
   test:
@@ -27,9 +32,6 @@ jobs:
         os: [macos-latest, ubuntu-latest]
 
     steps:
-      - name: Cancel Previous Runs
-        uses: styfle/cancel-workflow-action@0.12.1
-
       - name: Checkout
         uses: actions/checkout@v4
         with:
