### 2024-06-20 #6840 @stefanazzz
Hi,

yes that is good. I have no further suggestion,
thanks,

On Wed, Jun 19, 2024 at 8:57 AM Dongdong Tian ***@***.***>
wrote:

> @stefanazzz <https://github.com/stefanazzz> Thanks for your review.
>
> chekbox 'performance' and 'automated tests' not applicable.
>
> The HinetPy package contains many tests at
> https://github.com/seisman/HinetPy/tree/main/tests. These tests cover 84%
> of the HinetPy source codes and are automatically executed by GitHub
> Actions for every change made to the main branch (see
> https://github.com/seisman/HinetPy/actions/runs/9577487328 for the latest
> run of the tests).
>
> checkbox 'community guidelines' yes to question (2)
>
> In response to YuYifan2000's reviews (seisman/HinetPy#117
> <https://github.com/seisman/HinetPy/issues/117>), I've added a section to
> the README file in the main branch:
>
> Feedback and contributions are welcome! Please feel free to open an issue
> or pull request if you have any suggestions or would like to contribute a
> feature. For additional information or specific questions, please open an
> issue directly.
>
> I feel these sentences clarify that users should open issues or pull
> requests if they want to (1) Contribute to the software, 2) Report issues
> or problems with the software 3) Seek support.
>
> Is there anything else that you suggest to add?
>
