### 2024-06-20 #6840 @elbeejay
@stefanazzz I'd like to ask that you re-read and check off the remaining checkboxes on your checklist if possible. For the cases that are not applicable (such as the "Human and animal research" box, please check it off acknowledging that you saw the box). We like to get reviewers to check off their full checklists just as a measure of accountability for JOSS if we ever need to go back and revisit reviews. Thanks!
