### 2024-06-21 #6840 @seisman
@stefanazzz The HinetPy package already supports downloading F-net data. The F-net network ID is `0103` and the data are in WIN32 format. If you prefer to download F-net data in MSEED/SEED/SAC format from the F-net website (https://www.fnet.bosai.go.jp/), you can try the [FnetPy](https://github.com/seisman/FnetPy) package. The FnetPy is in its early stages and is not actively maintained, so it may not work as expected. Feel free to open issues in that repository and I'll find time fix them if any.
