### 2024-06-21 #6840 @elbeejay
@stefanazzz looking for your sign-off on the remaining two checkboxes in your checklist for the items "Automated tests" and "Community guidelines" - thanks!
