### 2024-06-25 #6840 @elbeejay
Many thanks to our reviewers @zhong-yy, @YuYifan2000, and @stefanazzz for taking the time to review this submission and provide their feedback.

@seisman - I will now go over the paper from an editorial perspective, and then provide you with a list of items to complete for me so that we can square away the metadata prior to getting this HinetPy paper published with JOSS.
