### 2024-06-25 #6840 @elbeejay
@seisman I have a few minor comments on the paper text:

Line 75: Consider rewording to "The key features of HinetPy are:" or "Below the key features of HinetPy are listed:" or something along those lines.

Lines 80-81: Swap the square bracket "[" and "]" for parentheses "(" and ")"

Line 84: Consider rephrasing to something like "Uses multithreading to reduce download and processing times" or "Speeds up the downloading and processing workflow via the use of multithreading"
