### 2024-06-25 #6840 @elbeejay
@seisman if you could do the "additional author tasks" in the comment above and let me know what the outcomes are, specifically I need to know if the version has changed and what the archived DOI is. Thanks!
