### 2024-06-25 8be109e
Remove the draft pdf workflow
