### 2024-06-25 #6840 @seisman
Many thanks to @elbeejay and the three reviewers for handling and reviewing this submission.

> Line 75: Consider rewording to "The key features of HinetPy are:" or "Below the key features of HinetPy are listed:" or something along those lines.
>
> Lines 80-81: Swap the square bracket "[" and "]" for parentheses "(" and ")"
>
> Line 84: Consider rephrasing to something like "Uses multithreading to reduce download and processing times" or "Speeds up the downloading and processing workflow via the use of multithreading"

I've made the changes in https://github.com/seisman/HinetPy/pull/120/commits/bd3bf768011a7f154b0c8b1d81d499e08a1ec011.

> Double check authors and affiliations (including ORCIDs)

Yes.

> Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.

[HinetPy 0.9.0](https://github.com/seisman/HinetPy/releases/tag/0.9.0) is the release for the paper.

> Archive the release on Zenodo/figshare/etc and post the DOI here.
Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.

Archived on Zenodo at https://doi.org/10.5281/zenodo.12523911

> Make sure that the license listed for the archive is the same as the software license.

Yes, the MIT license.
