### 2024-06-25 #6840 @elbeejay
Awesome thanks @seisman - I've recommended that we accept and publish this paper in JOSS. An editor-in-chief will do one final review of the information and may have minor comments or suggestions.
