### 2024-06-25 5565081
Add the JOSS paper (#120)
