+++ paper/paper.md
@@ -1,0 +1,161 @@
+---
+title: 'HinetPy: A Python package for accessing and processing NIED Hi-net seismic data'
+tags:
+  - Python
+  - geophysics
+  - seismology
+authors:
+ - name: Dongdong Tian
+   orcid: 0000-0001-7967-1197
+   affiliation: 1
+affiliations:
+ - name: Hubei Subsurface Multi-scale Imaging Key Laboratory, School of Geophysics and Geomatics, China University of Geosciences, Wuhan 430074, China
+   index: 1
+date: 3 April 2024
+bibliography: paper.bib
+---
+
+# Summary
+
+HinetPy is a Python package designed for researchers working with seismic data from the
+National Research Institute for Earth Science and Disaster Resilience (NIED) Hi-net
+(High-sensitivity seismograph network) in Japan. The seismic network comprises approximately
+800 stations with high-quality seismic data. Accessing and processing the data can be
+challenging due to the limited functionality of the web UI and backend data server.
+Additionally, the seismic data is stored in a non-standard format, which adds an extra
+layer of complexity. HinetPy solves these challenges by offering a user-friendly interface
+for accessing seismic data from NIED Hi-net and converting it to commonly used data
+formats. It streamlines the workflow for seismologists, enabling them to more effectively
+utilize this valuable dataset.
+
+# Statement of need
+
+The National Research Institute for Earth Science and Disaster Resilience (NIED) operates
+and maintains NIED Hi-net, a nationwide high-sensitivity seismograph network in Japan.
+Since its establishment in October 2000, NIED Hi-net has grown to include approximately
+800 seismic stations equipped with three-component short-period seismometers [@Obara2005; @Okada2014].
+The NIED Hi-net website ([https://www.hinet.bosai.go.jp/](https://www.hinet.bosai.go.jp/))
+provides access to high-quality seismic data from 2004 onwards, including data from
+other seismic networks such as F-net, S-net, V-net, and more. The NIED Hi-net data has
+been widely used in various research from the study of earthquakes [e.g., @Ishii2005; @Peng2007]
+to the structure of the Earth's deep interior [e.g., @Niu2005; @Yee2014; @Tian2017].
+Despite the value of the data provided by NIED Hi-net, accessing and processing it can be challenging.
+
+## Challenges in accessing NIED Hi-net data
+
+The NIED Hi-net data is available for free through the NIED Hi-net website after users
+register for an account. However, accessing them can still be challenging. While most seismic
+data centers have transitioned to standard FDSN web services
+([https://www.fdsn.org/webservices/](https://www.fdsn.org/webservices/)) in recent years,
+allowing users to request seismic data using open-source tools like ObsPy [@ObsPy2015],
+NIED Hi-net has not yet upgraded to adopt these services. To request data
+from the NIED Hi-net website, users must log in and click many buttons manually.
+It is important to note that the NIED Hi-net website has limitations on the data
+size and length in a single request. Specifically, the record length of a single channel
+cannot exceed 60 minutes, and the total record length of
+all channels cannot exceed 12,000 minutes. Considering that NIED Hi-net comprises about 800 seismic
+stations and 2,400 channels (3 channels per station), the record length in a single
+request must not exceed 5 minutes. If users require 30 minutes of data, they must divide the
+time range into six subranges and submit six separate requests. The NIED Hi-net website
+does not allow users to post multiple data requests simultaneously. Therefore, to obtain
+the requested data, users must post a request, wait for the data to be prepared, and then
+post subsequent requests. After all data is ready, users must manually download the files
+and combine them into a single file. The entire process can be time-consuming and cumbersome.
+
+## Challenges in processing NIED Hi-net data
+
[…]
