### 2024-06-25 #6840 @kthyng
Congratulations on your new publication @seisman! Many thanks to @elbeejay and to reviewers @zhong-yy, @YuYifan2000, and @stefanazzz for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.
